Reviewed-on: http://learn.zone01oujda.ma/git/hkhlifi/01blog/pulls/1

+++ backend/src/main/java/com/dev/backend/dto/ProfileUserResponse.java
@@ -10,7 +10,7 @@ public record ProfileUserResponse(
         String joinDate,
         int followers,
         int following,
-        int posts,
+        long posts,
         boolean isOwner,
         boolean isFollowing) {
 }
